requirements.txt
web3==6.11.1
eth-account==0.9.0
eth-typing==3.5.1
eth-utils==2.3.1
fastapi==0.104.1
uvicorn==0.24.0
websockets==12.0
aiohttp==3.9.1

# Database and Caching
redis==5.0.1
sqlalchemy==2.0.23

# Monitoring and Security
prometheus-client==0.19.0
python-json-logger==2.0.7
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Utilities
pyyaml==6.0.1
python-dotenv==1.0.0

